refactor(toolruntime): split each streamer into focused files
