+++ toolruntime/chat_stream.go
@@ -674,277 +674,6 @@ func (s *chatStreamer) emitData(body map[string]any) {
 	s.flusher.Flush()
 }
 
-// finalize emits any remaining buffered citations, forwards client-owned
-// tool_calls the model asked the user to execute, writes the finish-reason
-// chunk and final usage chunk, fires the billing event, and closes the
-// stream with `[DONE]`.
-func (s *chatStreamer) finalize(
-	r *http.Request,
-	em *manager.EnclaveManager,
-	modelName string,
-	result chatIterationResult,
-	clientToolCalls []toolCall,
-) error {
-	// Defensive: in the degenerate case where upstream never emitted a
-	// single chunk before signaling done (and the pump therefore never
-	// ran ensureRoleEmitted), the finish chunk below would be the
-	// client's first visible delta; still emit role first to keep the
-	// protocol contract.
-	s.ensureRoleEmitted()
-	s.flushCitations()
-
-	if len(clientToolCalls) > 0 {
-		raw := rawToolCallsFromParsed(clientToolCalls, result.rawToolCalls)
-		s.writeChunk(map[string]any{
-			"choices": []any{
-				map[string]any{
-					"index": 0,
-					"delta": map[string]any{"tool_calls": raw},
-				},
-			},
-		})
-	}
-
-	finishReason := result.finishReason
-	if finishReason == "" {
-		if len(clientToolCalls) > 0 {
-			finishReason = "tool_calls"
-		} else {
-			finishReason = "stop"
-		}
-	}
-	s.writeChunk(map[string]any{
-		"choices": []any{
-			map[string]any{
-				"index":         0,
-				"delta":         map[string]any{},
-				"finish_reason": finishReason,
-			},
-		},
-	})
-
-	finalUsage := s.finalUsage()
-	if s.clientRequestedUsage && finalUsage != nil {
-		s.writeChunk(map[string]any{
-			"choices": []any{},
-			"usage":   finalUsage,
-		})
-	}
-
-	if s.writeErr == nil {
-		if _, err := io.WriteString(s.w, "data: [DONE]\n\n"); err != nil {
-			s.writeErr = err
-		} else {
-			s.flusher.Flush()
-		}
-	}
-
-	if s.usageMetricsRequested && finalUsage != nil {
-		s.w.Header().Set(manager.UsageMetricsResponseHeader, formatUsageHeader(usageFromRawMap(finalUsage)))
-	}
-	s.emitBillingEvent(r, em, modelName, finalUsage)
-	return s.writeErr
-}
-
-// finalUsage assembles the aggregated usage block across every upstream
-// iteration in the OpenAI Chat Completions shape. The underlying
-// accumulator already handles turns where upstream omitted the usage block
-// entirely by falling back to zero.
-func (s *chatStreamer) finalUsage() map[string]any {
-	usage := s.usageTotals.Usage()
-	if usage == nil {
-		return nil
-	}
-	return map[string]any{
-		"prompt_tokens":     usage.PromptTokens,
-		"completion_tokens": usage.CompletionTokens,
-		"total_tokens":      usage.TotalTokens,
-	}
-}
-
-// emitBillingEvent delegates to the shared billing emitter used by the
-// non-streaming path so streaming and non-streaming requests produce the
-// same billing-event shape, including Tinfoil-Enclave attribution and
-// request-id resolution from upstream headers.
-func (s *chatStreamer) emitBillingEvent(r *http.Request, em *manager.EnclaveManager, modelName string, usage map[string]any) {
-	if em == nil {
-		return
-	}
-	header := s.upstreamHeaders
-	if header == nil {
-		header = http.Header{}
-	}
-	response := &upstreamJSONResponse{
-		header: header,
-		body:   map[string]any{"usage": usage},
-	}
-	emitBillingEvent(em, r, response, modelName, true)
-}
-
-// terminateWithError is the single point at which a mid-stream upstream
-// failure is surfaced to the client. If no SSE has started yet we return
-// the upstream error to the caller so Handle turns it into a plain JSON
-// error response; otherwise we emit a top-level error data frame in the
-// shape OpenAI SDKs recognize mid-stream, followed by `[DONE]` so SDKs can
-// close their streams cleanly.
-func (s *chatStreamer) terminateWithError(r *http.Request, em *manager.EnclaveManager, modelName string, err error) error {
-	if !s.headersWritten {
-		return err
-	}
-	s.flushCitations()
-	s.emitData(map[string]any{"error": upstreamErrorPayload(err)})
-	if s.writeErr == nil {
-		if _, werr := io.WriteString(s.w, "data: [DONE]\n\n"); werr != nil {
-			s.writeErr = werr
-		} else {
-			s.flusher.Flush()
-		}
-	}
-	s.emitBillingEvent(r, em, modelName, s.finalUsage())
-	return nil
-}
-
-// upstreamErrorPayload extracts the most informative structured error
-// object we can show the client. If the underlying error is an
-// *upstreamError whose body is JSON with an `error` field, we return that
-// object verbatim (preserving the upstream's own `code`, `type`, `param`,
-// etc.). Otherwise we fall back to a minimal `{message, type}` envelope.
-func upstreamErrorPayload(err error) map[string]any {
-	if upErr, ok := err.(*upstreamError); ok && len(upErr.body) > 0 {
-		var parsed map[string]any
-		if json.Unmarshal(upErr.body, &parsed) == nil {
-			if inner, ok := parsed["error"].(map[string]any); ok {
-				return inner
-			}
-		}
-	}
-	return map[string]any{
-		"message": err.Error(),
-		"type":    "upstream_error",
-	}
-}
-
-// chatToolCallBuilder assembles OpenAI streaming tool_call deltas into the
-// final tool_call objects. OpenAI emits each function's name and id on
-// the first delta for a given index, and the arguments JSON in incremental
-// string fragments; we replay that protocol byte-for-byte.
-type chatToolCallBuilder struct {
-	entries []*chatToolCallEntry
-}
-
-type chatToolCallEntry struct {
-	index        int
-	id           string
-	toolType     string
-	functionName string
-	arguments    []byte
-}
-
-// ingest merges one tool_call delta list into the builder's running state.
-// It preserves the original wire-format fragment in s.raw so the assistant
-// message emitted on the next iteration carries bytewise-identical tool
-// calls to whatever upstream produced.
-func (b *chatToolCallBuilder) ingest(rawCalls []any) {
-	for _, item := range rawCalls {
-		m, _ := item.(map[string]any)
-		if m == nil {
-			continue
-		}
-		idx := int(numberValue(m["index"]))
-		for len(b.entries) <= idx {
-			b.entries = append(b.entries, &chatToolCallEntry{index: len(b.entries)})
-		}
-		entry := b.entries[idx]
-		if id := stringValue(m["id"]); id != "" {
-			entry.id = id
-		}
-		if t := stringValue(m["type"]); t != "" {
-			entry.toolType = t
-		}
-		if fn, ok := m["function"].(map[string]any); ok {
-			if name := stringValue(fn["name"]); name != "" {
-				entry.functionName = name
-			}
-			if args, ok := fn["arguments"].(string); ok && args != "" {
-				entry.arguments = append(entry.arguments, args...)
-			}
-		}
-	}
-}
-
-// toolCalls returns the assembled tool_call objects as the parsed router
-// type. Arguments are JSON-decoded when possible; malformed arguments are
-// preserved as an empty map so the loop can still surface the tool call
-// rather than silently drop it.
-func (b *chatToolCallBuilder) toolCalls() []toolCall {
-	out := make([]toolCall, 0, len(b.entries))
-	for _, entry := range b.entries {
-		if entry == nil || entry.functionName == "" {
-			continue
-		}
-		args := map[string]any{}
-		if len(entry.arguments) > 0 {
-			raw, _ := sanitizeToolCallArgumentsJSON(entry.arguments)
-			_ = json.Unmarshal(raw, &args)
-		}
-		out = append(out, toolCall{
-			id:        entry.id,
-			name:      entry.functionName,
-			arguments: args,
-		})
-	}
-	return out
-}
-
-// raw returns the tool_calls array in the wire shape OpenAI uses for
-// non-streaming completions. Used to rehydrate the assistant message we
-// prepend to the next iteration's messages array.
-func (b *chatToolCallBuilder) raw() []any {
-	calls := make([]any, 0, len(b.entries))
-	for _, entry := range b.entries {
-		if entry == nil || entry.functionName == "" {
-			continue
-		}
-		calls = append(calls, map[string]any{
-			"id":   entry.id,
-			"type": nonEmptyString(entry.toolType, "function"),
-			"function": map[string]any{
-				"name":      entry.functionName,
-				"arguments": string(entry.arguments),
-			},
-		})
-	}
-	return calls
-}
-
-// rawToolCallsFromParsed filters the streamed raw tool_calls down to the
-// client-owned ones and re-indexes them starting at zero so the final
-// stream chunk emits a delta.tool_calls array in the shape clients expect:
-// contiguous `index` values identifying positions in the client-visible
-// tool_calls list. The `id`, `type`, and `function` fields are preserved
-// byte-for-byte from what upstream produced.
-func rawToolCallsFromParsed(parsed []toolCall, raw []any) []any {
-	allowed := make(map[string]struct{}, len(parsed))
-	for _, call := range parsed {
-		allowed[call.id] = struct{}{}
-	}
-	filtered := make([]any, 0, len(parsed))
-	for _, item := range raw {
-		source, _ := item.(map[string]any)
-		if _, ok := allowed[stringValue(source["id"])]; !ok {
-			continue
-		}
-		reindexed := map[string]any{
-			"index":    len(filtered),
-			"id":       source["id"],
-			"type":     source["type"],
-			"function": source["function"],
-		}
-		filtered = append(filtered, reindexed)
-	}
-	return filtered
-}
-
 // nestedAnnotationsFromMatches converts the internal citationAnnotation
 // slice into the nested url_citation shape documented by OpenAI for Chat
 // Completions responses so SDKs and UIs that already consume the final
