+++ toolruntime/responses_stream.go
@@ -499,634 +499,3 @@ func (s *responsesStreamer) pumpUpstream(reader *sseReader, isFirst bool) (respo
 	}
 	return result, nil
 }
-
-// captureIdentity records the upstream-provided response id, creation
-// time and model on the first event that carries each one. Subsequent
-// iterations retain the first captured value per field so the client sees
-// one stable logical stream across every tool turn, and so the
-// synthesized response.created and response.completed events carry
-// upstream's real created_at instead of the router's wall clock.
-func (s *responsesStreamer) captureIdentity(event map[string]any) {
-	response, _ := event["response"].(map[string]any)
-	if response == nil {
-		return
-	}
-	if !s.upstreamIDCaptured {
-		if id := stringValue(response["id"]); id != "" {
-			s.responseID = id
-			s.upstreamIDCaptured = true
-		}
-	}
-	if s.createdAt == 0 {
-		if created := int64(numberValue(response["created_at"])); created > 0 {
-			s.createdAt = created
-		}
-	}
-	if s.model == "" {
-		if model := stringValue(response["model"]); model != "" {
-			s.model = model
-		}
-	}
-}
-
-// ensureResponseCreated issues the synthesized response.created event
-// using the streamer's stable identity fields, exactly once per logical
-// stream. Callers may invoke it from the first upstream `response.created`
-// frame (in which case the identity accessors return upstream's values)
-// or from finalize (defensive, for upstreams that never send one). The
-// accessors fall back to router-minted defaults only when strictly
-// needed.
-func (s *responsesStreamer) ensureResponseCreated() {
-	if s.responseCreatedEmitted {
-		return
-	}
-	s.responseCreatedEmitted = true
-	s.emitEvent("response.created", map[string]any{
-		"type": "response.created",
-		"response": map[string]any{
-			"id":         s.streamResponseID(),
-			"object":     "response",
-			"created_at": s.streamCreatedAt(),
-			"status":     "in_progress",
-			"model":      s.streamModel(),
-			"output":     []any{},
-		},
-	})
-}
-
-// handleOutputItemAdded forwards the event to the client after remapping
-// output_index to the streamer's monotonically increasing namespace, or
-// suppresses it entirely if the item represents a router-owned tool call
-// we will execute ourselves. Item accumulation happens on the matching
-// `_done` event so this handler does not need the iteration result.
-//
-// Client-owned function_call items are forwarded live so the calling SDK
-// can act on them (e.g., execute the tool on its side and post results).
-// Only tool names listed in s.ownedTools are suppressed; unknown tools
-// default to client-owned to match the non-streaming splitToolCalls
-// classification.
-func (s *responsesStreamer) handleOutputItemAdded(event map[string]any) {
-	item, _ := event["item"].(map[string]any)
-	upstreamIndex := int(numberValue(event["output_index"]))
-	if s.isRouterOwnedToolItem(item) {
-		// Router-owned tool call; intercept for later resolution and do
-		// not surface the intermediate tool invocation to the client.
-		// mcp_call is included because some upstreams model the same
-		// function invocation under the MCP output shape when tools are
-		// presented via the legacy MCP path, and we want both to be
-		// resolved identically.
-		s.suppressedItems[upstreamIndex] = struct{}{}
-		return
-	}
-	clientIndex := s.outputIndex
-	s.outputIndexMap[upstreamIndex] = clientIndex
-	s.outputIndex++
-	payload := cloneJSONMap(event)
-	payload["output_index"] = clientIndex
-	s.emitEvent("response.output_item.added", payload)
-}
-
-// isRouterOwnedToolItem reports whether the given output item is a tool
-// invocation this router will execute internally. Any function_call or
-// mcp_call whose name is in the configured ownedTools set is suppressed
-// from the client stream; everything else (including function_calls for
-// tools the caller declared themselves) is forwarded verbatim.
-func (s *responsesStreamer) isRouterOwnedToolItem(item map[string]any) bool {
-	if item == nil {
-		return false
-	}
-	itemType := stringValue(item["type"])
-	if itemType != "function_call" && itemType != "mcp_call" {
-		return false
-	}
-	if s.ownedTools == nil {
-		return false
-	}
-	_, ok := s.ownedTools[stringValue(item["name"])]
-	return ok
-}
-
-// handleOutputItemDone completes the client-facing view of an item and
-// appends the finished item to finalOutput so response.completed can echo
-// every output the model produced. For BOTH router-owned and client-owned
-// function calls the parsed arguments are recorded for the loop; when the
-// upstream `done` event lacks arguments (some servers ship them only via
-// the delta stream) we fall back to the per-index buffer accumulated from
-// `response.function_call_arguments.delta` frames. For client-owned
-// calls the rebuilt arguments string is also written back into the item
-// so the forwarded `output_item.done` event and the terminal
-// `response.completed.output` both carry a usable tool call.
-func (s *responsesStreamer) handleOutputItemDone(event map[string]any, result *responsesIterationResult) {
-	item, _ := event["item"].(map[string]any)
-	upstreamIndex := int(numberValue(event["output_index"]))
-	itemType := stringValue(item["type"])
-	if s.isRouterOwnedToolItem(item) {
-		arguments := parseArguments(item["arguments"])
-		if len(arguments) == 0 {
-			if buf, ok := s.functionCallArguments[upstreamIndex]; ok {
-				arguments = parseArguments(buf.String())
-				// Splice the reassembled JSON back into the item before
-				// it lands in result.outputItems. The outer loop replays
-				// those items into the next /v1/responses turn's input,
-				// and some upstream variants ship arguments only via
-				// delta frames -- without this splice the replayed item
-				// would carry an empty arguments string and the model
-				// would see its own prior tool call as argument-less.
-				if rawArgs := stringValue(item["arguments"]); rawArgs == "" {
-					item["arguments"] = buf.String()
-				}
-			}
-		}
-		delete(s.functionCallArguments, upstreamIndex)
-		result.outputItems = append(result.outputItems, item)
-		callID := stringValue(item["call_id"])
-		if callID == "" {
-			callID = stringValue(item["id"])
-		}
-		result.toolCalls = append(result.toolCalls, toolCall{
-			id:        callID,
-			name:      stringValue(item["name"]),
-			arguments: arguments,
-		})
-		return
-	}
-	if _, suppressed := s.suppressedItems[upstreamIndex]; suppressed {
-		return
-	}
-	// Client-owned function_calls: reassemble arguments from the delta
-	// buffer when the item itself ships with an empty arguments string,
-	// and splice the reassembled form back into the item so downstream
-	// consumers see a complete tool call.
-	if itemType == "function_call" || itemType == "mcp_call" {
-		if rawArgs := stringValue(item["arguments"]); rawArgs == "" {
-			if buf, ok := s.functionCallArguments[upstreamIndex]; ok && buf.Len() > 0 {
-				item["arguments"] = buf.String()
-			}
-		}
-		delete(s.functionCallArguments, upstreamIndex)
-	}
-	clientIndex, ok := s.outputIndexMap[upstreamIndex]
-	if !ok {
-		clientIndex = s.outputIndex
-		s.outputIndexMap[upstreamIndex] = clientIndex
-		s.outputIndex++
-	}
-	payload := cloneJSONMap(event)
-	payload["output_index"] = clientIndex
-	s.emitEvent("response.output_item.done", payload)
-	s.finalOutput = append(s.finalOutput, item)
-	result.outputItems = append(result.outputItems, item)
-	if itemType == "function_call" || itemType == "mcp_call" {
-		// Client-owned tool call: propagate to iteration result so
-		// splitToolCalls classifies it as a client tool call and the
-		// outer loop falls through to finalize rather than attempting to
-		// execute it internally.
-		callID := stringValue(item["call_id"])
-		if callID == "" {
-			callID = stringValue(item["id"])
-		}
-		result.toolCalls = append(result.toolCalls, toolCall{
-			id:        callID,
-			name:      stringValue(item["name"]),
-			arguments: parseArguments(item["arguments"]),
-		})
-	}
-}
-
-// handleFunctionCallArgumentsDelta accumulates partial JSON arguments for
-// a router-owned function_call item that the loop will resolve internally.
-// The buffered string is consumed when the matching
-// `response.output_item.done` event fires.
-func (s *responsesStreamer) handleFunctionCallArgumentsDelta(event map[string]any) {
-	upstreamIndex := int(numberValue(event["output_index"]))
-	delta := stringValue(event["delta"])
-	if delta == "" {
-		return
-	}
-	if s.functionCallArguments == nil {
-		s.functionCallArguments = map[int]*strings.Builder{}
-	}
-	buf, ok := s.functionCallArguments[upstreamIndex]
-	if !ok {
-		buf = &strings.Builder{}
-		s.functionCallArguments[upstreamIndex] = buf
-	}
-	buf.WriteString(delta)
-}
-
-// handleOutputTextDelta pushes the incoming text fragment through the
-// content's citation emitter so inline markdown links resolve to
-// annotation events, and forwards remaining text to the client as an
-// output_text.delta event.
-func (s *responsesStreamer) handleOutputTextDelta(event map[string]any) {
-	upstreamIndex := int(numberValue(event["output_index"]))
-	if _, suppressed := s.suppressedItems[upstreamIndex]; suppressed {
-		return
-	}
-	clientIndex, ok := s.outputIndexMap[upstreamIndex]
-	if !ok {
-		clientIndex = s.outputIndex
-		s.outputIndexMap[upstreamIndex] = clientIndex
-		s.outputIndex++
-	}
-	itemID := stringValue(event["item_id"])
-	contentIndex := int(numberValue(event["content_index"]))
-	key := itemContentKey{itemID: itemID, contentIndex: contentIndex}
-	emitter, exists := s.emitters[key]
-	if !exists {
-		emitter = newCitationEmitter(s.citations)
-		s.emitters[key] = emitter
-	}
-	delta := stringValue(event["delta"])
-	contentChunk, annotations := emitter.push(delta)
-	if contentChunk != "" {
-		s.emitEvent("response.output_text.delta", map[string]any{
-			"type":          "response.output_text.delta",
-			"output_index":  clientIndex,
-			"item_id":       itemID,
-			"content_index": contentIndex,
-			"delta":         contentChunk,
-		})
-	}
-	s.emitAnnotationEvents(annotations, clientIndex, itemID, contentIndex)
-}
-
-// handleOutputTextDone flushes any content still held in the emitter's
-// trailing buffer so the user sees every rune before the item closes, then
-// forwards the done event with identity rewritten.
-func (s *responsesStreamer) handleOutputTextDone(event map[string]any) {
-	upstreamIndex := int(numberValue(event["output_index"]))
-	if _, suppressed := s.suppressedItems[upstreamIndex]; suppressed {
-		return
-	}
-	clientIndex, ok := s.outputIndexMap[upstreamIndex]
-	if !ok {
-		clientIndex = s.outputIndex
-		s.outputIndexMap[upstreamIndex] = clientIndex
-		s.outputIndex++
-	}
-	itemID := stringValue(event["item_id"])
-	contentIndex := int(numberValue(event["content_index"]))
-	key := itemContentKey{itemID: itemID, contentIndex: contentIndex}
-	if emitter, ok := s.emitters[key]; ok {
-		contentChunk, annotations := emitter.flush()
-		if contentChunk != "" {
-			s.emitEvent("response.output_text.delta", map[string]any{
-				"type":          "response.output_text.delta",
-				"output_index":  clientIndex,
-				"item_id":       itemID,
-				"content_index": contentIndex,
-				"delta":         contentChunk,
-			})
-		}
-		s.emitAnnotationEvents(annotations, clientIndex, itemID, contentIndex)
-		delete(s.emitters, key)
-	}
-	payload := cloneJSONMap(event)
-	payload["output_index"] = clientIndex
-	s.emitEvent("response.output_text.done", payload)
-}
-
-// emitAnnotationEvents turns resolved link matches into the documented
-// Responses annotation event shape: `response.output_text.annotation.added`
-// with a flat url_citation attached to the item/content coordinates the
-// matching text lives at. The `annotation_index` is a monotonic cursor
-// per (item_id, content_index) pair so clients can use it to dedupe or
-// order annotations across the full text stream, not just within the
-// single push/flush batch that happened to resolve them.
-func (s *responsesStreamer) emitAnnotationEvents(matches []citationAnnotation, outputIndex int, itemID string, contentIndex int) {
-	if len(matches) == 0 {
-		return
-	}
-	if s.annotationCounts == nil {
-		s.annotationCounts = map[itemContentKey]int{}
-	}
-	key := itemContentKey{itemID: itemID, contentIndex: contentIndex}
-	for _, match := range matches {
-		annotation := map[string]any{
-			"type":        "url_citation",
-			"url":         match.source.url,
-			"start_index": match.startIndex,
-			"end_index":   match.endIndex,
-		}
-		if match.source.title != "" {
-			annotation["title"] = match.source.title
-		}
-		annotationIndex := s.annotationCounts[key]
-		s.annotationCounts[key] = annotationIndex + 1
-		s.emitEvent("response.output_text.annotation.added", map[string]any{
-			"type":             "response.output_text.annotation.added",
-			"output_index":     outputIndex,
-			"item_id":          itemID,
-			"content_index":    contentIndex,
-			"annotation_index": annotationIndex,
-			"annotation":       annotation,
-		})
-	}
-}
-
-// forwardEvent relays an upstream event the streamer does not specially
-// handle, rewriting output_index when present so the client always sees
-// the streamer's canonical monotonic namespace.
-func (s *responsesStreamer) forwardEvent(eventType string, event map[string]any) {
-	if rawIndex, ok := event["output_index"]; ok {
-		upstreamIndex := int(numberValue(rawIndex))
-		if _, suppressed := s.suppressedItems[upstreamIndex]; suppressed {
-			return
-		}
-		if clientIndex, mapped := s.outputIndexMap[upstreamIndex]; mapped {
-			event = cloneJSONMap(event)
-			event["output_index"] = clientIndex
-		}
-	}
-	s.emitEvent(eventType, event)
-}
-
-// executeTool invokes the MCP session and surfaces router-owned web
-// search progress as the spec-conformant Responses streaming events
-// documented by OpenAI:
-//
-//   - response.output_item.added (item type web_search_call, status in_progress)
-//   - response.web_search_call.in_progress
-//   - response.web_search_call.searching (search only; fetch omits this phase)
-//   - response.web_search_call.completed (on success) OR no dedicated event on failure
-//   - response.output_item.done (terminal status completed or failed)
-//
-// The stream is always fully spec-conformant on the Responses path; no
-// opt-in marker channel is used. Clients that want to distinguish a
-// safety-filter block from a generic failure can inspect the non-spec
-// `status` strings on the terminal `web_search_call` output item carried
-// in `response.completed.output`, which is collapsed onto `failed` at
-// the envelope level but preserved on the record for tooling.
-func (s *responsesStreamer) executeTool(ctx context.Context, session *mcp.ClientSession, call toolCall) (string, error) {
-	switch call.name {
-	case "search":
-		id := "ws_" + uuid.NewString()
-		action := map[string]any{"type": "search", "query": stringValue(call.arguments["query"])}
-		outputIndex := s.openWebSearchCallItem(id, action)
-		s.emitWebSearchCallPhase("response.web_search_call.in_progress", id, outputIndex)
-		s.emitWebSearchCallPhase("response.web_search_call.searching", id, outputIndex)
-		output, err := callTool(ctx, session, call.name, call.arguments, s.citations)
-		if err != nil {
-			reason := publicToolErrorReason(call.name, err)
-			status := "failed"
-			if reason == blockedToolErrorReason {
-				status = "blocked"
-			}
-			s.closeWebSearchCallItem(id, outputIndex, action, status, reason)
-			return "", err
-		}
-		s.emitWebSearchCallPhase("response.web_search_call.completed", id, outputIndex)
-		s.closeWebSearchCallItem(id, outputIndex, action, "completed", "")
-		return output, nil
-	case "fetch":
-		urls := fetchArgumentURLs(call.arguments)
-		if len(urls) == 0 {
-			return callTool(ctx, session, call.name, call.arguments, s.citations)
-		}
-		fetchIDs := make([]string, len(urls))
-		fetchIndices := make([]int, len(urls))
-		fetchActions := make([]map[string]any, len(urls))
-		for i, url := range urls {
-			fetchIDs[i] = "ws_" + uuid.NewString()
-			fetchActions[i] = map[string]any{"type": "open_page", "url": url}
-			fetchIndices[i] = s.openWebSearchCallItem(fetchIDs[i], fetchActions[i])
-			s.emitWebSearchCallPhase("response.web_search_call.in_progress", fetchIDs[i], fetchIndices[i])
-		}
-		output, err := callTool(ctx, session, call.name, call.arguments, s.citations)
-		if err != nil {
-			reason := publicToolErrorReason(call.name, err)
-			status := "failed"
-			if reason == blockedToolErrorReason {
-				status = "blocked"
-			}
-			for i := range urls {
-				s.closeWebSearchCallItem(fetchIDs[i], fetchIndices[i], fetchActions[i], status, reason)
-			}
-			return "", err
-		}
-		for i := range urls {
-			s.emitWebSearchCallPhase("response.web_search_call.completed", fetchIDs[i], fetchIndices[i])
-			s.closeWebSearchCallItem(fetchIDs[i], fetchIndices[i], fetchActions[i], "completed", "")
-		}
-		return output, nil
-	default:
-		return callTool(ctx, session, call.name, call.arguments, s.citations)
-	}
-}
-
-// openWebSearchCallItem emits response.output_item.added for a
-// router-owned web_search_call item and returns the output_index the
-// client will see for it. The item is NOT recorded into s.finalOutput
-// because attachResponsesCitations rebuilds the terminal snapshot from
-// citations.toolCalls; the live event here is solely for the client's
-// progress UI.
-func (s *responsesStreamer) openWebSearchCallItem(id string, action map[string]any) int {
-	outputIndex := s.outputIndex
-	s.outputIndex++
-	s.emitEvent("response.output_item.added", map[string]any{
-		"type":         "response.output_item.added",
-		"output_index": outputIndex,
-		"item": map[string]any{
-			"id":     id,
-			"type":   "web_search_call",
-			"status": "in_progress",
-			"action": action,
-		},
-	})
-	return outputIndex
-}
-
-// emitWebSearchCallPhase emits a progress envelope for a web_search_call
-// item in the shape OpenAI's spec defines: {item_id, output_index,
-// sequence_number, type}. Only response.web_search_call.{in_progress,
-// searching, completed} are spec-defined phase events; failures surface
-// solely through the terminal response.output_item.done envelope.
-func (s *responsesStreamer) emitWebSearchCallPhase(eventType, itemID string, outputIndex int) {
-	s.emitEvent(eventType, map[string]any{
-		"type":         eventType,
-		"item_id":      itemID,
-		"output_index": outputIndex,
-	})
-}
-
-// closeWebSearchCallItem emits response.output_item.done for a
-// router-owned web_search_call item with the given terminal status. The
-// spec defines the envelope status enum as {in_progress, searching,
-// completed, failed}, so any internal `blocked` value is collapsed
-// inside webSearchCallEvent and the unfiltered router status (plus any
-// error code) rides on the `_tinfoil` sidecar for clients that want to
-// render a distinct affordance for safety-filter blocks.
-func (s *responsesStreamer) closeWebSearchCallItem(id string, outputIndex int, action map[string]any, status, errorCode string) {
-	item := webSearchCallEvent(id, status, errorCode, action)
-	s.emitEvent("response.output_item.done", map[string]any{
-		"type":         "response.output_item.done",
-		"output_index": outputIndex,
-		"item":         item,
-	})
-}
-
-// finalize writes the aggregated `response.completed` event with every
-// output item the model produced plus the accumulated usage, fires the
-// billing event, and closes the SSE stream.
-func (s *responsesStreamer) finalize(r *http.Request, em *manager.EnclaveManager, modelName string, result responsesIterationResult) error {
-	// Cover the degenerate case where upstream never emitted a
-	// response.created frame (or emitted it on an iteration other than
-	// the first, which we ignore). The client still needs to see exactly
-	// one response.created before any response.completed.
-	s.ensureResponseCreated()
-
-	// By the time finalize runs, every text part has received its
-	// matching output_text.done, which already drained the emitter for
-	// that (item_id, content_index). Anything still in s.emitters is
-	// necessarily empty; clearing the map ensures we do not leak state
-	// across requests if the streamer is ever reused.
-	s.emitters = map[itemContentKey]*citationEmitter{}
-
-	usage := result.usage
-	if usage == nil {
-		usage = s.aggregatedUsage
-	}
-	totals := s.usageTotals.Usage()
-	if totals != nil {
-		usage = map[string]any{
-			"input_tokens":  totals.PromptTokens,
-			"output_tokens": totals.CompletionTokens,
-			"total_tokens":  totals.TotalTokens,
-		}
-	}
-	// Build the final `output` array. Run it through attachResponsesCitations
-	// on a synthetic body so the final snapshot carries the same normalized
-	// text and the same flat url_citation annotations the non-streaming path
-	// produces, plus the prepended web_search_call items for every
-	// router-owned tool call. Clients that consume response.completed
-	// instead of (or in addition to) the live delta stream must see a
-	// terminal payload that is byte-for-byte comparable to the
-	// non-streaming response.
-	completedBody := map[string]any{"output": append([]any{}, s.finalOutput...)}
-	// The Responses streaming path surfaces web_search progress via the
-	// spec-defined `response.web_search_call.*` events fired live from
-	// executeTool and via the prepended `web_search_call` output items
-	// attachResponsesCitations injects here. No tinfoil-event marker
-	// channel rides on this path -- the Responses stream is always
-	// fully OpenAI-spec-conformant.
-	attachResponsesCitations(completedBody, s.citations, s.includeActionSources)
-	finalOutput, _ := completedBody["output"].([]any)
-	completed := map[string]any{
-		"id":         s.streamResponseID(),
-		"object":     "response",
-		"created_at": s.streamCreatedAt(),
-		"status":     "completed",
-		"model":      s.streamModel(),
-		"output":     finalOutput,
-	}
-	if usage != nil {
-		completed["usage"] = usage
-	}
-	s.emitEvent("response.completed", map[string]any{
-		"type":     "response.completed",
-		"response": completed,
-	})
-
-	if s.usageMetricsRequested && usage != nil {
-		s.w.Header().Set(manager.UsageMetricsResponseHeader, formatUsageHeader(usageFromRawMap(usage)))
-	}
-	s.emitBillingEvent(r, em, modelName, usage)
-	return s.writeErr
-}
-
-// emitBillingEvent delegates to the shared billing emitter used by the
-// non-streaming path so streaming and non-streaming requests produce the
-// same billing-event shape, including Tinfoil-Enclave attribution taken
-// from the most recent upstream response headers.
-func (s *responsesStreamer) emitBillingEvent(r *http.Request, em *manager.EnclaveManager, modelName string, usage map[string]any) {
-	if em == nil {
-		return
-	}
-	header := s.upstreamHeaders
-	if header == nil {
-		header = http.Header{}
-	}
-	response := &upstreamJSONResponse{
-		header: header,
-		body:   map[string]any{"usage": usage},
-	}
-	emitBillingEvent(em, r, response, modelName, true)
-}
-
-// terminateWithError surfaces a mid-stream upstream failure to the client.
-// If no SSE has started, the error bubbles up so Handle can turn it into a
-// plain JSON error response; otherwise the streamer emits a synthesized
-// response.failed event and the billing pipeline records every token
-// accumulated across every successful turn so far, not just the failing
-// one. Multi-turn tool loops would otherwise undercount billing when a
-// later iteration fails.
-func (s *responsesStreamer) terminateWithError(r *http.Request, em *manager.EnclaveManager, modelName string, err error) error {
-	if !s.headersWritten {
-		return err
-	}
-	s.ensureResponseCreated()
-	s.emitEvent("response.failed", map[string]any{
-		"type": "response.failed",
-		"response": map[string]any{
-			"id":         s.streamResponseID(),
-			"object":     "response",
-			"created_at": s.streamCreatedAt(),
-			"status":     "failed",
-			"model":      s.streamModel(),
-			"output":     s.finalOutput,
-			"error":      upstreamErrorPayload(err),
-		},
-	})
-	s.emitBillingEvent(r, em, modelName, s.totalsBillingUsage())
-	return nil
-}
-
-// totalsBillingUsage returns the aggregated Responses-shaped usage block
-// spanning every iteration the stream has consumed, falling back to the
-// most recent single-turn usage block if no totals were recorded.
-func (s *responsesStreamer) totalsBillingUsage() map[string]any {
-	if totals := s.usageTotals.Usage(); totals != nil {
-		return map[string]any{
-			"input_tokens":  totals.PromptTokens,
-			"output_tokens": totals.CompletionTokens,
-			"total_tokens":  totals.TotalTokens,
-		}
-	}
-	return s.aggregatedUsage
-}
-
-// extractResponsesUsage pulls the usage block out of a response.completed
-// / response.failed / response.incomplete event, where it lives under the
-// nested `response` object per the documented Responses API event shape.
-func extractResponsesUsage(event map[string]any) map[string]any {
-	response, _ := event["response"].(map[string]any)
-	if response == nil {
-		return nil
-	}
-	usage, _ := response["usage"].(map[string]any)
-	return usage
-}
-
-// extractResponsesErrorMessage pulls a human-readable reason out of a
-// response.failed or response.incomplete event. Upstream may place the
-// detail under response.error.message or response.incomplete_details;
-// callers treat an empty return as "no extra detail available" and fall
-// back to the event type.
-func extractResponsesErrorMessage(event map[string]any) string {
-	response, _ := event["response"].(map[string]any)
-	if response == nil {
-		return ""
-	}
-	if errObj, ok := response["error"].(map[string]any); ok {
-		if message := stringValue(errObj["message"]); message != "" {
-			return message
-		}
-	}
-	if details, ok := response["incomplete_details"].(map[string]any); ok {
-		if reason := stringValue(details["reason"]); reason != "" {
-			return reason
-		}
-	}
-	return ""
-}
